Renamed core/elizabeth.py to providers.py
