+++ elizabeth/core/__init__.py
@@ -1,4 +1,4 @@
-from .elizabeth import (
+from .providers import (
     Address, Business, ClothingSizes,
     Code, Datetime, Development, File,
     Food, Hardware, Internet, Network,
