+++ elizabeth/core/interdata/address.py
@@ -296,7 +296,6 @@ POSTAL_CODE_FMT = {
     'jp': '###-####',
     'hu': '####',
     'is': '###',
-    'jp': '###-####',
     'ko': '#####',
     'nl': '1### @@',
     'no': '####',
