+++ tests/test_data/test_clothing_sizes.py
@@ -2,7 +2,7 @@
 
 from unittest import TestCase
 
-from elizabeth.core.elizabeth import ClothingSizes
+from elizabeth.core.providers import ClothingSizes
 
 
 class ClothingSizesTest(TestCase):
